\name{covB}
\alias{covB}
\title{Computes part of the variance covariance matrix for fixed effects}
\usage{
  covB(X, fits)
}
\arguments{
  \item{X}{design matrix}

  \item{fits}{fits}
}
\value{
  a matrix
}
\description{
  Computes part of the variance covariance matrix for fixed
  effects
}
\author{
  Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}

