\name{distBetween2xySets}
\alias{distBetween2xySets}
\title{distance between two matrices, data.frames, or sp SpatialPoints
distance between two matrices, data.frames, or sp SpatialPoints}
\usage{
  distBetween2xySets(xy1, xy2)
}
\arguments{
  \item{xy1}{internal}

  \item{xy2}{internal}
}
\value{
  matrix of distances
}
\description{
  distance between two matrices, data.frames, or sp
  SpatialPoints distance between two matrices, data.frames,
  or sp SpatialPoints
}
\author{
  Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}

