\name{plot.spCountSamp}
\alias{plot.spCountSamp}
\title{Default plotting method for spCountSamp objects.  Maps the prediction surface.}
\usage{
  plot.spCountSamp(ob, cex.leglab = 1, cex.main = 1,
    main = "Study Area")
}
\arguments{
  \item{ob}{fitted model of class spCountSamp}

  \item{cex.leglab}{character expansion for legend symbols
  and characters}

  \item{main}{same as for par}

  \item{cex.main}{character expansion for title main}
}
\value{
  a plot. The map of the outline, sample plots, and
  prediction surface.
}
\description{
  Default plotting method for spCountSamp objects.  Maps
  the prediction surface.
}
\author{
  Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}

