% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spglasamp.R
\name{plot.spglasamp}
\alias{plot.spglasamp}
\title{Plots the fitted spglasamp object}
\usage{
\method{plot}{spglasamp}(
  obj,
  cex.pred = 0.3,
  cex.knots = 2,
  pch.coarse = 3,
  pch.fine = 4,
  lwd.knots = 3,
  nclass = 9
)
}
\arguments{
\item{obj}{a spglasamp object}
}
\value{
a plot
}
\description{
Plots the fitted spglasamp object
}
\author{
Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}
