% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSummary_original.R
\name{SimSummary.original}
\alias{SimSummary.original}
\title{Summary statistics from simulated data
Summary statistics from simulated data}
\usage{
SimSummary.original(
  StoreSims,
  RootName,
  ODType = NA,
  failCutEst = NA,
  failCutSE = NA,
  failCutOD = NA
)
}
\arguments{
\item{StoreSims}{isppifac results stored for simulated data}

\item{RootName}{Root name for a column in StoreSims, either "SRS", "SPP5",
"Spp10", "Spp20", or "Spp30"}

\item{ODType}{Overdispersion type; either "Trad" or "Perc or Vary"}

\item{StdErrType}{Either "Spp"" or "GLMM""}
}
\value{
data.frame of summary statistics
}
\description{
Summary statistics from simulated data
Summary statistics from simulated data
}
\author{
Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}
