\name{summary.spCountSamp}
\alias{summary.spCountSamp}
\title{Default summary method for spCountSamp objects
Default summary method for spCountSamp objects}
\usage{
  \method{summary}{spCountSamp} (ob)
}
\arguments{
  \item{ob}{spCountSamp objects}
}
\value{
  list of summary statistics, including the estimate, a
  data.frame of various standard error estimates, a
  data.frame of proportions of the variance components, and
  a data.frame of the range parameters.
}
\description{
  Default summary method for spCountSamp objects Default
  summary method for spCountSamp objects
}
\author{
  Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}

