% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spCountSamp.R
\name{summary.spglasamp}
\alias{summary.spglasamp}
\title{Default summary method for spCountSamp objects
Default summary method for spCountSamp objects}
\usage{
\method{summary}{spglasamp}(ob)
}
\arguments{
\item{ob}{spglasamp objects}
}
\value{
list of summary statistics, including the estimate, a data.frame
of various standard error estimates, a data.frame of proportions of the
variance components, and a data.frame of the range parameters.
}
\description{
Default summary method for spCountSamp objects
Default summary method for spCountSamp objects
}
\author{
Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}
